Add a number of permissions-related classes and changes.
A Student now has a permissions property of type StudentPermissions.

The PermissionsWindow is functioning, but can't be reached yet until the Roster is done.

Permissionlevel.py is an enum class that has the three whiteboard permission levels.
There is also an enum package now which everyone can use if they need it.