Return GenericFailure for failed connection
I wanted to handle this in the __connectCentral method, but it's not possible
without storing a reference to the callback somewhere. That degrades rapidly
if the user, let's say preses submit twice quickly.

Return a GenericFailure to the user with a conneciton failed method if
connecting to the central server failed.