Complete student-central server process in join
This is the beginning of join-presentation. First the student requests to join a
presentation. The central server validates the request and responds to the
student. That is the summary of what this commit does.

Next up: the central server responded with connection info for the student
to connect to the presenter.. connect to the presenter now.