Send student HTML
The student now no longer has to have a copy of the HTML on his machine.
The student is automatically sent a copy of whatever html the presenter has
open.

Also, eclass.initialData is added. This is useful if you want to send students
any information upon logging in (eg what slide they should be on,
what the initial messages in the forum are.. etc) do a "git grep" for
"initialData" and read the comments so you know how to use it.

The place it's available for use at (as the comments state) is view/JoinPresentation
see the example there in the callback() method