Remove __joinedClass property on student connections if..
If the student attempts to join a presentaiton and succeeds on all validation
centeral server side, but then fails due to validataion on the presenter's side
(eg the student is kicked).. the presenter needs to inform the central server
that validation failed.

This commit does so. And by doing so gets rid of some annoying errors with kicking
a student. Also, this commit fixes some lamesauce stuff where
trying to join a class that isn't hosted resulted in the connection being closed.