Allow same student to reconnect during presenation
There was an error which used to occur when a student rejoined a presentation,
and then the presenter tried to send something to the student (eg sync slide on
second join during same presentation).

This happened because we do checking to make sure that students only have one
connection to the presenter at any given time, and I forgot to remove the
connnections as students leave the presentation.